progress/common.h -> _options.h; Status is no longer common
