+++ c/options.h
@@ -7,9 +7,9 @@
 //------------------------------------------------------------------------------
 #ifndef dt_OPTIONS_h
 #define dt_OPTIONS_h
-#include <functional>  // std::function
-#include "python/ext_type.h"
-#include "python/obj.h"
+#include <functional>         // std::function
+namespace py { class oobj; }  // forward-declare py::oobj
+
 namespace dt {
 
 
@@ -22,41 +22,8 @@ void use_options_store(py::oobj options);
 
 py::oobj get_option(const char* name);
 
-
-}
-//------------------------------------------------------------------------------
-// config_option
-//------------------------------------------------------------------------------
-namespace py {
+void init_config_option(void* module);
 
 
-struct config_option : public PyObject {
-  std::function<oobj()> getter;
-  std::function<void(oobj)> setter;
-  oobj name;
-  oobj default_value;
-  oobj docstring;
-  size_t : 64;
-
-  class Type : public ExtType<config_option> {
-    public:
-      static PKArgs args___init__;
-      static const char* classname();
-      static const char* classdoc();
-      static bool is_subclassable();
-      static void init_methods_and_getsets(Methods&, GetSetters&);
-  };
-
-  void m__init__(PKArgs&);
-  void m__dealloc__();
-  oobj get(const PKArgs&);
-  void set(const PKArgs&);
-
-  oobj get_name() const;
-  oobj get_doc() const;
-  oobj get_default() const;
-};
-
-
-}  // namespace py
+}  // namespace dt
 #endif
